git-svn-id: https://anrl.googlecode.com/svn/trunk@336 3ecfba0c-028a-11de-ada4-5153b8187bf2

+++ AirNavigationRaceLive/AnrlInterfaces/IGPSPoint.cs
@@ -7,8 +7,8 @@ namespace AnrlInterfaces
 {
     public interface IGPSPoint
     {
-        Decimal Longitude { get; }
-        Decimal Latitude { get; }
-        Decimal Altitude { get; }
+        Double Longitude { get; }
+        Double Latitude { get; }
+        Double Altitude { get; }
     }
 }
